CXP-1353: externalisation of OpenAppButton from ConnectedAppContainer
